change cache time

+++ app/controllers/green_mill_controller.rb
@@ -14,7 +14,7 @@ class GreenMillController < ActionController::API
     else
     	timeDiff = ((Time.now.to_f - File.ctime(fileLoc).to_f) / 3600.0)
 			puts "Time difference: ", timeDiff
-      if timeDiff > 3
+      if timeDiff > 3 && timeDiff < 0.5
         puts "Reloaded cache"
 				File.delete(fileLoc)
 		  	result = build_json_from_html
